Add files via upload
This R program file contains the functions to cache a matrix and to solve for the inverse of a matrix. The code in the file is for work that I completed as a part of the R Programming Course available on Course offered by John Hopkins University's Dr. Peng. For more details on this course, here's a link: https://www.coursera.org/learn/r-programming?specialization=data-science-foundations-r